# apps/backend/src/main/java/ma/farmsense/dto/poultry/CreateFlockRequest.java
package ma.farmsense.dto.poultry;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import ma.farmsense.entity.FlockPurpose;

import java.time.LocalDate;
import java.util.UUID;

@Data @Builder @NoArgsConstructor @AllArgsConstructor
public class CreateFlockRequest {

    @NotBlank
    @Size(max = 255)
    private String name;

    @Size(max = 255)
    private String nameAr;

    @Size(max = 255)
    private String nameEn;

    @Size(max = 255)
    private String breed;

    @NotNull
    @Min(1)
    private Integer birdCount;

    @NotNull
    private FlockPurpose purpose;

    private LocalDate startDate;
    private UUID supplierId;
    private String source;
    private String notes;
}